requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
pandas==2.2.0
openpyxl==3.1.2
Pillow==10.2.0
reportlab==4.0.9
numpy==1.26.3
gunicorn==21.2.0
